Hello plugin now uses nixysa to generate NPRuntime bindings.
Words cannot describe how inherently cool this is.  Input a static lib and
an annotated header file (idl), and you'll get a plugin that works on
FF 3.0, 3.5, Safari 4, Chrome 3, 4 on OS X, Windows, and Linux (where
applicable) which exposes the lib to JavaScript using the interface specified
in the idl.  And it's the same build command on every OS.

Still some cleanup needed here and there, but hey, it works for HelloWorld.