Builds with latest dev version of SCons.
This is a horrible horrible mess.  Something has gone wrong with SCons
dependency tracking on autogenerated cc files that then get compiled into
a VariantDir.  I had previously worked around this by saying the file in the
VariantDir depends on the file not in the VariantDir, but SCons has gotten
more clever and now ignores that (the file depending on itself really doesn't
make sense).  Unfortunately it hadn't gotten clever enough to recognize that
the file in the VariantDir relies on the generator being run.

The new workaround is that the file in the VariantDir copies its sources and
builder from the generated file node, and then is marked as a duplicate of
the generated file node.  This results in SCons running the generator and then
creating a symlink from the VariantDir to the generated file, which is good
because with the new file being explicitly created with Environment.File,
SCons is failing to recognize that it's in a VariantDir and thus passing the
path inside the VariantDir to the compiler instead of the file's real path.

tl;dr version: Ugh.