Massive cleanup to Build(), including one last tweak to its dependency tracking.
So, I'm an idiot and tried to solve a dependency problem with a dependency
solution instead of solving the more obvious file system problem with a file
system solution.  All of the crazy "deal with borked dependencies" code has
been replaced by "copy the file from <file> to build/<cfg>/<file>, which works
every time, on every OS, with every version of SCons.

To be fair it took a while to figure exactly why SCons was breaking in this
case, and all the crazy things I did helped me figure that out.  Also being
fair to SCons, it's not really "breaking", this behavior is by design.  It
just doesn't make sense for the kind of multistage build process I'm doing.

See the comment that starts at Config.py:101 for more information.

In addition to that I refactored some more preparation/cleanup code from
Build() into helper functions, so the code that does important things is more
apparent.